make large file a folder
